fix: A safer way to manage pages cache
