+++ jest-e2e.config.js
@@ -1,8 +1,8 @@
 module.exports = {
     preset: 'jest-puppeteer',
     // Uncomment the `testMatch` option when we want to run a specific test case
-    testMatch: ['<rootDir>/packages/default-layout/__tests__/Bookmarks.e2e.ts'],
-    // testRegex: ['(/__tests__/.*|(\\.|/)(e2e))\\.ts$'],
+    // testMatch: ['<rootDir>/packages/default-layout/__tests__/Bookmarks.e2e.ts'],
+    testRegex: ['(/__tests__/.*|(\\.|/)(e2e))\\.ts$'],
     transform: {
         '^.+\\.tsx?$': 'ts-jest',
     },
