+++ packages/core/src/annotations/Link.tsx
@@ -8,8 +8,8 @@
 
 import * as React from 'react';
 
-import { usePages } from '../hooks/usePages';
 import { SpecialZoomLevel } from '../structs/SpecialZoomLevel';
+import { getDestination } from '../utils/managePages';
 import { Annotation } from './Annotation';
 import type { PdfJs } from '../types/PdfJs';
 
@@ -21,13 +21,11 @@ export const Link: React.FC<{
     onExecuteNamedAction(action: string): void;
     onJumpToDest(pageIndex: number, bottomOffset: number, leftOffset: number, scaleTo: number | SpecialZoomLevel): void;
 }> = ({ annotation, doc, page, viewport, onExecuteNamedAction, onJumpToDest }) => {
-    const { getDestination } = usePages(doc);
-
     const link = (e: React.MouseEvent): void => {
         e.preventDefault();
         annotation.action
             ? onExecuteNamedAction(annotation.action)
-            : getDestination(annotation.dest).then((target) => {
+            : getDestination(doc, annotation.dest).then((target) => {
                   const { pageIndex, bottomOffset, leftOffset, scaleTo } = target;
                   onJumpToDest(pageIndex, bottomOffset, leftOffset, scaleTo);
               });
