+++ packages/core/src/layouts/PageSizeCalculator.tsx
@@ -9,8 +9,8 @@
 import * as React from 'react';
 
 import { Spinner } from '../components/Spinner';
-import { usePages } from '../hooks/usePages';
 import { SpecialZoomLevel } from '../structs/SpecialZoomLevel';
+import { getPage } from '../utils/managePages';
 import { decrease } from '../zoom/zoomingLevel';
 import { calculateScale } from './calculateScale';
 import type { PageSize } from '../types/PageSize';
@@ -21,7 +21,6 @@ export const PageSizeCalculator: React.FC<{
     doc: PdfJs.PdfDocument;
     render(pageSize: PageSize): React.ReactElement;
 }> = ({ defaultScale, doc, render }) => {
-    const { getPage } = usePages(doc);
     const pagesRef = React.useRef<HTMLDivElement | null>(null);
     const [pageSize, setPageSize] = React.useState<PageSize>({
         pageHeight: 0,
@@ -30,7 +29,7 @@ export const PageSizeCalculator: React.FC<{
     });
 
     React.useEffect(() => {
-        getPage(0).then((pdfPage) => {
+        getPage(doc, 0).then((pdfPage) => {
             const viewport = pdfPage.getViewport({ scale: 1 });
             const w = viewport.width;
             const h = viewport.height;
