+++ packages/thumbnail/src/ThumbnailContainer.tsx
@@ -7,7 +7,7 @@
  */
 
 import * as React from 'react';
-import { useIntersectionObserver, usePages } from '@react-pdf-viewer/core';
+import { getPage, useIntersectionObserver } from '@react-pdf-viewer/core';
 import type { PdfJs, VisibilityChanged } from '@react-pdf-viewer/core';
 
 import { SpinnerContext } from './SpinnerContext';
@@ -32,7 +32,6 @@ export const ThumbnailContainer: React.FC<{
     rotation: number;
     onActive(pageIndex: number): void;
 }> = ({ doc, isActive, pageHeight, pageIndex, pageWidth, rotation, onActive }) => {
-    const { getPage } = usePages(doc);
     const [pageSize, setPageSize] = React.useState<PageState>({
         height: pageHeight,
         isCalculated: false,
@@ -64,7 +63,7 @@ export const ThumbnailContainer: React.FC<{
     const onVisibilityChanged = (params: VisibilityChanged): void => {
         visibilityRef.current = params;
         if (params.isVisible && !isCalculated) {
-            getPage(pageIndex).then((pdfPage) => {
+            getPage(doc, pageIndex).then((pdfPage) => {
                 const viewport = pdfPage.getViewport({ scale: 1 });
 
                 setPageSize({
